labs/LangChain_103/rag_ollama_llava_drive/requirements.txt
langchain-core 
langchain-community 
streamlit 
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
